requirements.txt
aiofiles==23.1.0
cryptography==41.0.7
pysocks==1.7.1
click==8.1.7
pyyaml==6.0.1
colorlog==6.7.0
requests==2.31.0
requests[socks]==2.31.0 